refactor: #56/매장별 메뉴 다건 조회 권한 설정 추가 및 수정
