+++ src/main/java/com/sparta/delivery_app/domain/user/entity/UserStatus.java
@@ -1,9 +1,12 @@
 package com.sparta.delivery_app.domain.user.entity;
 
 import com.sparta.delivery_app.common.exception.errorcode.UserErrorCode;
+import com.sparta.delivery_app.common.globalcustomexception.UnableOpenStoreException;
 import com.sparta.delivery_app.common.globalcustomexception.UserStatusException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+
+import static com.sparta.delivery_app.common.exception.errorcode.UserErrorCode.NOT_AUTHORITY_TO_REGISTER;
 
 @Getter
 @RequiredArgsConstructor
@@ -16,6 +19,12 @@ public enum UserStatus {
     public static void checkUserStatus(UserStatus userRole) {
         if (userRole == DISABLE) {
             throw new UserStatusException(UserErrorCode.USER_STATUS_DISABLE);
+        }
+    }
+
+    public static void checkManagerEnable(User user) {
+        if (!user.getUserStatus().equals(UserStatus.ENABLE)) {
+            throw new UnableOpenStoreException(NOT_AUTHORITY_TO_REGISTER);
         }
     }
 }
